requirements.txt
# Dev
black>=22.6.0
mypy>=0.961
pylint>=2.14.4
pytest>=7.1.2

# Project
duneapi==6.0.0
dune-client>=0.0.6
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
web3>=5.30.0
psycopg2-binary>=2.9.3
SQLAlchemy==1.4.41
pandas==1.5.0
click==8.1.3
